refactor: マイク処理をカスタムフックに分離
